Enable initial position param in kinova xacros
The `initial_positions` parameter is already defined in the `load_arm`
and `kortex_ros2_control` macros, it is just unable to be used when
loading a kinova robot because it is not passed through in the
`load_robot` macro.

These changes add the `initial_positions` param in the `load_robot`
macro and pass it through to the `load_arm` macro so that a kinova robot
can be loaded with parameterized initial positions.