Use internal comm param in gripper macros
Add `use_internal_bus_gripper_comm` param back into `load_gripper` macro
and use it to determine whether or not to launch a ros2_control instance
for the gripper. We expect `use_internal_bus_gripper_comm` to be false
if we are running the gripper in simulation or the gripper is
connected to the PC via USB rather than through the internal kinova
bus. In those cases we need to launch a ros2_control instance to control
the gripper.

Remove `gripper_include_ros2_control` param from macros since we're
using the `use_internal_bus_gripper_comm` param and the simulation
params to determine if we need to launch a ros2_control instance for the
gripper.

Update the kortex_description readme to reflect these changes.